requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
spacy==3.7.2

# Run after install: python -m spacy download en_core_web_sm